requirements.txt
mysql-connector-python>=8.0.33
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
schedule>=1.2.0
python-dateutil>=2.8.2
